updated nppesquerier tests to follow integration test paradigms
