+++ endpointmanager/pkg/endpointmanager/postgresql/npiorganizationstore_test.go
@@ -1,20 +1,19 @@
+// +build integration
+
 package postgresql
 
 import (
 	"testing"
 
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/endpointmanager"
-	"github.com/spf13/viper"
+	th "github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/testhelper"
 )
 
 func Test_DeleteAllNPIOrganizations(t *testing.T) {
-	var err error
+	teardown, _ := th.IntegrationDBTestSetup(t, store.DB)
+	defer teardown(t, store.DB)
 
-	store, err := NewStore(viper.GetString("dbhost"), viper.GetInt("dbport"), viper.GetString("dbuser"), viper.GetString("dbpass"), viper.GetString("dbname"), viper.GetString("dbsslmode"))
-	if err != nil {
-		t.Errorf("Error creating Store type: %s", err.Error())
-	}
-	defer store.Close()
+	var err error
 
 	var npio1 = &endpointmanager.NPIOrganization{
 		ID:            1,
@@ -97,13 +96,10 @@ func Test_DeleteAllNPIOrganizations(t *testing.T) {
 }
 
 func Test_PersistNPIOrganization(t *testing.T) {
-	var err error
+	teardown, _ := th.IntegrationDBTestSetup(t, store.DB)
+	defer teardown(t, store.DB)
 
-	store, err := NewStore(viper.GetString("dbhost"), viper.GetInt("dbport"), viper.GetString("dbuser"), viper.GetString("dbpass"), viper.GetString("dbname"), viper.GetString("dbsslmode"))
-	if err != nil {
-		t.Errorf("Error creating Store type: %s", err.Error())
-	}
-	defer store.Close()
+	var err error
 
 	var npio1 = &endpointmanager.NPIOrganization{
 		ID:            1,
